requirements.txt
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.3
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mss==6.0.0
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
pure-python-adb==0.3.0.dev0
pycodestyle==2.6.0
pylint==2.6.0
pytesseract==0.3.6
six==1.15.0
toml==0.10.1
wrapt==1.12.1
